app/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
click==8.1.7
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
toml==0.10.2
typing_extensions==4.12.2
uvicorn==0.22.0